rename from Zensu to Telecine
All the monitoring stuff was removed a while ago. It makes more sense
to use this repository to build rpc stuff, then build zensu on top of
that.

The rename to Telecine reflects that. Telecine is "the process of
transferring film to video", so it sort of makes sense for an rpc
framework (being mostly concerned with format impedance). Plus it
continues the silly film references!