add a Serializer class
This takes care of annotating primitives with type information so they
can be sent to the correct classes on the other end. It supports
multiple backends and an options hash to be passed down the chain.

If a registered name cannot be found for the annotated name in the
representation, the primitive is returned unaltered. Future options
might allow raising an exception instead.

Also rename tc_load and tc_dump to from_primitive and as_primitive
respectively.